añado validación

+++ src/main/java/org/iesvdm/pedidos_ventas/dao/UtilDAO.java
@@ -24,7 +24,7 @@ public class UtilDAO {
                         rs.getString("CO.nombre"),
                         rs.getString("CO.apellido1"),
                         rs.getString("CO.apellido2"),
-                        rs.getDouble("CO.comisión")
+                        rs.getBigDecimal("CO.comisión")
                 )
         );
     }
@@ -43,6 +43,6 @@ public class UtilDAO {
                 , rs.getString("nombre")
                 , rs.getString("apellido1")
                 , rs.getString("apellido2")
-                , rs.getDouble("comisión"));
+                , rs.getBigDecimal("comisión"));
     }
 }
